Extract and fix more query tests
