+++ src/lib.rs
@@ -27,7 +27,9 @@ mod unittest_public_metadata;
 mod unittest_queries;
 mod unittest_reveal;
 mod unittest_transfer;
+mod unittest_query_all_nft_info;
 mod unittest_query_all_tokens;
+mod unittest_query_approved_for_all;
 mod unittest_query_contract_config;
 mod unittest_query_contract_info;
 mod unittest_query_minters;
@@ -35,8 +37,11 @@ mod unittest_query_nft_dossier;
 mod unittest_query_nft_info;
 mod unittest_query_num_tokens;
 mod unittest_query_owner_of;
+mod unittest_query_private_metadata;
+mod unittest_query_token_approvals;
 mod unittest_query_tokens;
 mod unittest_query_unwrapped;
+mod unittest_query_verify_transfer_approval;
 mod unittest_register_receive_nft;
 mod unittest_send;
 mod unittest_set_global_approval;
